requirements.txt
elasticsearch
fastapi
isort
flake8
python-dotenv
pytest
responses
uvicorn